# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Family Recipes",
      "description": "Recipes from the whole tribe",
      "version": "1.0.0"
    },
    "host": "localhost:3000",
    "schemes": [
      "http",
      "https"
    ],
    "produces": [
      "application/json",
      "application/xml"
    ],
    "consumes": [
      "application/json",
      "application/xml"
    ],
...
    "paths": {
      "/users": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get all users",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Post a user",
          "description": "",
          "operationId": "postUser",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "OK"
            },
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/users/{userId}": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get user by id",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "description": "User Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "Users"
          ],
          "summary": "Update an existing user",
          "description": "",
          "operationId": "updateUser",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "description": "User Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            },
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be modified",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "User not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        },
        "delete": {
          "tags": [
            "Users"
          ],
          "summary": "Delete an existing user",
          "description": "",
          "operationId": "deleteUser",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "description": "User Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "User not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        }
      },
...
      "/recipes": {
        "get": {
          "tags": [
            "Recipes"
          ],
          "summary": "Get all recipes",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "Recipes"
          ],
          "summary": "Post a new recipe",
          "description": "",
          "operationId": "postRecipe",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Recipe object that needs to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Recipe"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "OK"
            },
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/recipes/{recipeId}": {
        "get": {
          "tags": [
            "Recipes"
          ],
          "summary": "Get recipe by id",
          "parameters": [
            {
              "in": "path",
              "name": "recipeId",
              "description": "Recipe Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "Recipes"
          ],
          "summary": "Update an existing recipe",
          "description": "",
          "operationId": "updateRecipe",
          "parameters": [
            {
              "in": "path",
              "name": "recipeId",
              "description": "Recipe Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            },
            {
              "in": "body",
              "name": "body",
              "description": "Recipe object that needs to be modified",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Recipe"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "User not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        },
        "delete": {
          "tags": [
            "Recipes"
          ],
          "summary": "Delete an existing recipe",
          "description": "",
          "operationId": "deleteRecipe",
          "parameters": [
            {
              "in": "path",
              "name": "recipeId",
              "description": "Recipe Id",
              "required": true,
              "type": "integer",
              "minimum": 1
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "User not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Recipe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "category_id": {
            "type": "number"
          },
          "sub_category_id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "cooking_time": {
            "type": "number"
          },
          "preparation_time": {
            "type": "string"
          },
          "share_nb": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email_active": {
            "type": "number"
          },
          "role": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        }
      }
    }
  }